refactor: move ComponentType to common
