+++ src/libecalc/dto/base.py
@@ -3,35 +3,8 @@ from enum import Enum
 
 from pydantic import BaseModel, ConfigDict
 
+from libecalc.common.component_type import ComponentType
 from libecalc.common.string.string_utils import to_camel_case
-
-
-class ComponentType(str, Enum):
-    ASSET = "ASSET"
-    INSTALLATION = "INSTALLATION"
-    GENERATOR_SET = "GENERATOR_SET"
-
-    CONSUMER_SYSTEM_V2 = "CONSUMER_SYSTEM@v2"
-    COMPRESSOR_SYSTEM = "COMPRESSOR_SYSTEM"
-    PUMP_SYSTEM = "PUMP_SYSTEM"
-    COMPRESSOR = "COMPRESSOR"
-    COMPRESSOR_V2 = "COMPRESSOR@v2"
-    PUMP = "PUMP"
-    PUMP_V2 = "PUMP@v2"
-    GENERIC = "GENERIC"
-    # TURBINE = "TURBINE"
-    VENTING_EMITTER = "VENTING_EMITTER"
-    TRAIN_V2 = "TRAIN@V2"
-
-    def __lt__(self, other: "ComponentType"):  # type: ignore[override]
-        if self == other:
-            return False
-        # the following works because the order of elements in the definition is preserved
-        for elem in ComponentType:
-            if self == elem:
-                return True
-            elif other == elem:
-                return False
 
 
 class ConsumerUserDefinedCategoryType(str, Enum):
