+++ src/libecalc/dto/node_info.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel
 
 from libecalc.common.component_info.component_level import ComponentLevel
-from libecalc.dto.base import ComponentType
+from libecalc.common.component_type import ComponentType
 from libecalc.dto.utils.validators import ComponentNameStr
