+++ src/libecalc/fixtures/cases/all_energy_usage_models/all_models_dto.py
@@ -3,9 +3,9 @@ from datetime import datetime
 import pytest
 
 from libecalc import dto
+from libecalc.common.component_type import ComponentType
 from libecalc.common.utils.rates import RateType
 from libecalc.dto.base import (
-    ComponentType,
     ConsumerUserDefinedCategoryType,
     InstallationUserDefinedCategoryType,
 )
