+++ src/libecalc/fixtures/cases/consumer_system_v2/consumer_system_v2_dto.py
@@ -3,6 +3,7 @@ from datetime import datetime
 import pytest
 
 from libecalc import dto
+from libecalc.common.component_type import ComponentType
 from libecalc.common.string.string_utils import generate_id
 from libecalc.common.units import Unit
 from libecalc.common.utils.rates import RateType
@@ -16,7 +17,6 @@ from libecalc.dto import (
     PumpSystemPump,
 )
 from libecalc.dto.base import (
-    ComponentType,
     ConsumerUserDefinedCategoryType,
     FuelTypeUserDefinedCategoryType,
     InstallationUserDefinedCategoryType,
