+++ src/libecalc/fixtures/cases/consumer_with_time_slots_models.py
@@ -4,9 +4,9 @@ from datetime import datetime
 import pytest
 
 from libecalc import dto
+from libecalc.common.component_type import ComponentType
 from libecalc.common.variables import VariablesMap
 from libecalc.dto.base import (
-    ComponentType,
     ConsumerUserDefinedCategoryType,
     InstallationUserDefinedCategoryType,
 )
