+++ src/libecalc/presentation/exporter/aggregators.py
@@ -2,9 +2,9 @@ import abc
 from typing import List
 
 from libecalc.application.graph_result import GraphResult
+from libecalc.common.component_type import ComponentType
 from libecalc.common.logger import logger
 from libecalc.common.time_utils import Frequency
-from libecalc.dto.base import ComponentType
 from libecalc.presentation.exporter.appliers import Applier
 from libecalc.presentation.exporter.dto.dtos import GroupedQueryResult, QueryResult
