+++ src/libecalc/presentation/yaml/yaml_types/components/yaml_compressor.py
@@ -4,8 +4,8 @@ from typing import Dict, Literal, Optional, Union
 from pydantic import ConfigDict, Field
 
 from libecalc import dto
+from libecalc.common.component_type import ComponentType
 from libecalc.common.time_utils import Period, define_time_model_for_period
-from libecalc.dto.base import ComponentType
 from libecalc.dto.components import CompressorComponent
 from libecalc.dto.types import ConsumptionType
 from libecalc.expression import Expression
