+++ src/libecalc/presentation/yaml/yaml_types/components/yaml_pump.py
@@ -4,8 +4,8 @@ from typing import Dict, Literal, Optional
 from pydantic import ConfigDict, Field
 
 from libecalc import dto
+from libecalc.common.component_type import ComponentType
 from libecalc.common.time_utils import Period, define_time_model_for_period
-from libecalc.dto.base import ComponentType
 from libecalc.dto.components import PumpComponent
 from libecalc.dto.types import ConsumptionType
 from libecalc.expression import Expression
