+++ src/tests/libecalc/core/consumers/conftest.py
@@ -3,9 +3,9 @@ from datetime import datetime
 import pytest
 
 from libecalc import dto
+from libecalc.common.component_type import ComponentType
 from libecalc.common.utils.rates import RateType
 from libecalc.common.variables import VariablesMap
-from libecalc.dto.base import ComponentType
 from libecalc.expression import Expression
