+++ src/tests/libecalc/core/consumers/consumer_function/test_direct_expression_consumer_function.py
@@ -5,13 +5,13 @@ import pytest
 
 import libecalc.common.utils.rates
 from libecalc import dto
+from libecalc.common.component_type import ComponentType
 from libecalc.common.temporal_model import TemporalModel
 from libecalc.common.variables import VariablesMap
 from libecalc.core.consumers.legacy_consumer.component import Consumer
 from libecalc.core.consumers.legacy_consumer.consumer_function.direct_expression_consumer_function import (
     DirectExpressionConsumerFunction,
 )
-from libecalc.dto.base import ComponentType
 from libecalc.dto.types import ConsumptionType
 from libecalc.expression import Expression
