+++ src/tests/libecalc/dto/test_electricity_consumer.py
@@ -4,7 +4,7 @@ import pytest
 from pydantic import ValidationError
 
 from libecalc import dto
-from libecalc.dto.base import ComponentType
+from libecalc.common.component_type import ComponentType
 from libecalc.dto.types import EnergyUsageType
 from libecalc.expression import Expression
